pyproject.toml
[tool.poetry]
name = "haystack-playground"
version = "0.1.0"
description = ""
authors = ["Pao-Sheng Wang <paooap.oappao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
haystack-ai = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
